+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 # Changelog
- - Fixed some random bugs
+ - Added Stylized Popups
+ - Simplified and fixed `basic_dark`, `red` and `orange` themes
+ - Changed translation's way of working (can now get translation id `category/translation`)
+ - Can now load media from an url ("Netwok flux")
+ - Fixed the "no sound/video tracks menus" in Firefox
+ - Added an "offline" page
+ - Added a favicon
+ - Added a launcher (`run.py`)
+ - Added an update software (`update.py` and `updater.py`)
