+++ README.md
@@ -1,4 +1,4 @@
-# Web Media Player v.1.4
+# Web Media Player v.1.5
 My own little HTML/CSS/JS media player.
 
 [![@samuellouf](https://img.shields.io/badge/author-SamuelLouf-lightgray.svg)](https://github.com/samuellouf/WebMediaPlayer)
@@ -10,3 +10,19 @@ Source repo on [GitHub](https://github.com/samuellouf/WebMediaPlayer), and run i
 
 Web Media Player is a [VLC](https://www.videolan.org/)-like multimedia player.
 It supports several basic features (Play/Pause, Fullscreen, Picture in Picture...) but it also supports more features like color themes.
+
+## Usage
+
+### Online
+To use WebMediaPlayer online, just open "[https://samuellouf.github.io/WebMediaPlayer/player](https://samuellouf.github.io/WebMediaPlayer/player)".
+
+### Offline
+To use WebMediaPlayer offline, download "[https://github.com/samuellouf/WebMediaPlayer/archive/refs/heads/1.5.zip](https://github.com/samuellouf/WebMediaPlayer/archive/refs/heads/1.5.zip)".
+
+#### With Python
+Execute the file `run.py`.
+
+#### Without Python
+Open `player/index.html` in your browser.
+> [!IMPORTANT]
+> You will not be able to update WebMediaPlayer automatically without running it using python.
