+++ player/api/hasPythonAccess
@@ -1,0 +1,3 @@
+{
+  "hasPython": false
+}
