+++ player/offline.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Offline</title>
+  <style>
+    body{
+      font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif, Arial;
+    }
+  </style>
+</head>
+<body>
+  <h1 align="center">Offline</h1>
+  <br><br>
+  <p>WebMediaPlayer is <b>offline</b>.</p>
+</body>
+</html>
