+++ player/themes/basic_dark.css
@@ -1,136 +1,12 @@
-body{
-  background-color: #555555;
-  color: white;
-  transition-duration: 1s;
-}
-
-/* Context menu (right-click menu) */
-.cm_container ul{
-  background-color: #2d2d2d;
-  box-shadow: 0 0 5px #000000;
-}
-
-.cm_container li{
-  color: #ccc;
-}
-
-.cm_container li:hover{
-  background-color: #242424;
-}
-
-.cm_container li.cm_divider{
-  border-bottom: 1px solid #aaa;
-}
-
-.cm_container li[disabled=""]{
-  color: #777;
-}
-
-div.controls {
-  background-color: #f5f5f580;
-}
-
-div.audio.fake h1{
-  background-color: #f0efef;
-  color: rgb(0, 60, 255);
-}
-
-div.loading {
-  background-color: black;
-  color: white;
-}
-
-div.loading div.content{
-  color: white;
-}
-
-div.filename {
-  color: white;
-}
-
-:root {
-  --button-color: #ddd;
-}
-
-.speed-dropdown button:hover {
-  background-color: #d3e8ff;
-}
-
-:root {
-  --select-border: #777;
-  --select-focus: blue;
-}
-
-/* Apps Window */
-.appwindow {
-  background-color: #777;
-  border: 1px solid #555;
-}
-
-.appwindowheader {
-  background-color: #555;
-  color: #ffffff;
-}
-
-.appwindow select {
-  background-color: #fff;
-  background-image: linear-gradient(to top, #f9f9f9, #fff 33%);
-}
-
-.appwindow select[disabled] {
-  background-color: #eee;
-  background-image: linear-gradient(to top, #ddd, #eee 33%);
-}
-.appwindow {
-  background-color: #777;
-}
-
-.appwindow .appwindowheader {
-  background-color: #323232;
-}
-
-.appwindow .resizer{
-  color: #777;
-}
-
-/* Player */
-path.w, text.w{
-  fill: white;
-}
-
-path.ww, text.ww{
-  stroke: white;
-}
-
-g.w{
-  stroke: white;
-}
-
-/* Buttons */
-[poseas="button"]{
-  background-color: #646464;
-  border: #bcbcbc solid 2px;
-}
-
-[poseas="button"]:hover{
-  background-color: #6a9acf;
-  border: #4a7db3 solid 2px;
-}
-
-[poseas="button"]:active{
-  background-color: #b5bfe7;
-  border: #2559b3 solid 2px;
-}
-
 :root{
   --color: black;
   --background-color: #cecbcb;
-  --context-menu: #dbdbdb;
+  --context-menu: #777777;
   --context-menu-shadow: #626262;
-  --context-menu-hovered: #b0b0b0;
+  --context-menu-hovered: #242424;
   --context-menu-divider: #aaa;
   --context-menu-disabled: #777;
-  --controls: #f5f5f580;
+  --controls: #7d7d7d80;
   --audio-background-color: #f0efef;
   --audio-color: #003cff;
   --button-color: #ddd;
@@ -139,20 +15,20 @@ g.w{
   --loading-background-color: black;
   --loading-color: white;
   --filename-color: white;
-  --speed-dropdown-background-color: #f5f5f5;
+  --speed-dropdown-background-color: #d3e8ff;
   --appwindow-color: #fff;
-  --appwindow-backgroud-color: #efefef;
-  --appwindow-border: #d3d3d3;
-  --appwindow-header-background-color: #dcdcdc;
+  --appwindow-backgroud-color: #777;
+  --appwindow-border: #555;
+  --appwindow-header-background-color: #323232;
   --appwindow-header-color: white;
   --select-background-color: #fff;
   --select-background-image: linear-gradient(to top, #f9f9f9, #fff 33%);
   --select-disabled-background-color: #eee;
   --select-disabled-background-image: linear-gradient(to top, #ddd, #eee 33%);
-  --player-icons-colors: black;
-  --button-bg: #f0f0f0;
+  --player-icons-colors: white;
+  --button-bg: #646464;
   --button-border: #bcbcbc;
-  --hover-button-bg: #d3e8ff;
+  --hover-button-bg: #6a9acf;
   --hover-button-border: #4a7db3;
   --active-button-bg: #b5bfe7;
   --active-button-border: #2559b3;
@@ -163,7 +39,7 @@ g.w{
   --speed-dropdown-shadow: rgba(0, 0, 0, 0.1);
   --cc-menu-slider-bg: #ccc;
   --cc-menu-slider-before: white;
-  --cc-menu-bg: #f5f5f5;
+  --cc-menu-bg: #606060;
   --cc-menu-input-slider: #2196F3;
   --close-button: red;
   --tracks-bg: #000000a4;
@@ -171,7 +47,6 @@ g.w{
   --table-border: rgb(140 140 140);
   --thead-bg: rgb(228 240 245);
   --td-border: rgb(160 160 160);
-  --speed-dropdown-bg:#d3e8ff;
   --button-color: #ddd;
   --select-border: #777;
   --select-focus: blue;
