+++ player/themes/basic_light.css
@@ -47,7 +47,6 @@
   --table-border: rgb(140 140 140);
   --thead-bg: rgb(228 240 245);
   --td-border: rgb(160 160 160);
-  --speed-dropdown-bg:#d3e8ff;
   --button-color: #ddd;
   --select-border: #777;
   --select-focus: blue;
