+++ player/themes/orange.css
@@ -8,7 +8,10 @@
   --context-menu-disabled: #632e02;
   --controls: #f5f5f580;
   --audio-background-color: #f0efef;
-  --audio-color: #ff9600;
+  --audio-color: orange;
+  --button-color: #ddd;
+  --select-border: #777;
+  --select-focus: orange;
   --loading-background-color: black;
   --loading-color: white;
   --filename-color: white;
@@ -17,8 +20,9 @@
   --appwindow-backgroud-color: #ed6700;
   --appwindow-border: #d45100;
   --appwindow-header-background-color: #dd6700;
-  --select-background-color: #ff9600;
-  --select-background-image: linear-gradient(to top, #ff9600, #ff7100 33%);
+  --appwindow-header-color: white;
+  --select-background-color: orange;
+  --select-background-image: linear-gradient(to top, orange, #ff7100 33%);
   --select-disabled-background-color: #eee;
   --select-disabled-background-image: linear-gradient(to top, #ddd, #eee 33%);
   --player-icons-colors: black;
@@ -43,7 +47,6 @@
   --table-border: rgb(140 140 140);
   --thead-bg: rgb(228 240 245);
   --td-border: rgb(160 160 160);
-  --speed-dropdown-bg:#d3e8ff;
   --button-color: #ddd;
   --select-border: #777;
   --select-focus: orange;
