+++ updates.json
@@ -1,3 +1,3 @@
 {
-  "lastest_version": "1.4.0"
+  "lastest_version": "1.5.0"
 }
